TorusSDK class (WIP): * Added AbstractTorusSDK class for common code between web & native * TorusSDK web & native versions extend the Abstract; overriding initialize & constructor
